pyproject.toml
[tool.poetry]
name = "agendamento-back-v2"
version = "0.1.0"
description = ""
authors = ["Ismael Rodrigo <ismaelbrasil02@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
anyio="3.6.1"
bcrypt="3.2.2"
cffi="1.15.1"
click="8.1.3"
colorama="0.4.5"
cryptography="37.0.4"
ecdsa="0.18.0"
fastapi="0.78.0"
greenlet="1.1.2"
h11="0.13.0"
idna="3.3"
Mako="1.2.1"
MarkupSafe="2.1.1"
passlib="1.7.4"
pyasn1="0.4.8"
pycparser="2.21"
pydantic="1.9.1"
pymongo={extras = ["srv"], version = "^4.2.0"}
python-decouple="3.6"
python-jose="3.3.0"
rsa="4.8"
six="1.16.0"
sniffio="1.2.0"
starlette="0.19.1"
typing_extensions="4.3.0"
uvicorn="0.18.2"
python-multipart = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = 'src.scripts.run:dev'
